refactor(plugins): colocate unit tests with their modules
Per the test rules, manifest.rs and store.rs now carry their own
#[cfg(test)] modules; shared fixtures (temp roots, plugin folders,
in-memory zips) move to test_support.rs; and tests.rs shrinks to the
connected tests only: the Tauri command surface through a mock app and
the InstalledPlugin wire shape.
