+++ src-tauri/src/commands/plugins/mod.rs
@@ -35,6 +35,8 @@ pub struct InstalledPlugin {
 mod manifest;
 mod store;
 #[cfg(test)]
+mod test_support;
+#[cfg(test)]
 mod tests;
 
 use store::{install_into, install_package, read_asset_from, scan_plugins_root, uninstall_from};
